requirements.txt
SQLAlchemy>=2.0,<3.0
psycopg[binary]>=3.2,<4.0
python-dotenv>=1.0,<2.0
alembic>=1.13,<2.0
fastapi>=0.116,<1.0
uvicorn>=0.34,<1.0
jinja2>=3.1,<4.0
python-multipart>=0.0.20,<1.0
itsdangerous>=2.2,<3.0
Authlib>=1.3,<2.0
httpx>=0.27,<1.0
